Create Code in Arduino IDE
We First connected the DHT 11 Sensor on the Arduino Uno and then connected the Arduino Uno to the computer using an USB cable and installed Arduino IDE in the computer. Then We started the IDE and wrote the code in the IDE and Compiled it and it gave us Heat Index, the Temperature and the Humidity of the room we were in.